refactor(database): cleanup interface imports
