+++ modules/database/src/interfaces/index.ts
@@ -1,4 +1,8 @@
+export * from './ConduitBuiltRoute';
+export * from './ConduitDatabaseSchema';
 export * from './CustomEndpoint';
+export * from './CustomEndpointsQuery';
 export * from './DeclaredSchema';
 export * from './DeclaredSchemaExtension';
 export * from './SchemaAdapter';
+export * from './Sort';
